Use separate slugs for Organizations and Locations
If Location belongs to Organization, and a location's slug is composed of both the location name and the organization name, mongoid-slug can't keep track of changes to an organization's name, so a new slug won't get generated. This was the wrong approach. It's best to provide a separate slug for each and let web clients use that to construct their own URL paths.