Don't include results for generic email domains when searching by domain
Currently, the admin app authorizes users by comparing the domain name of their email address to the domain names of a location's emails and/or URLs. If a user's email address is one of the generic email services, like gmail, aol, hotmail, etc., then the search used to return all locations that have those generic email addresses.

This commit fixes that by specifying two different parameters that the admin app should use: "email" for finding an exact match for the entire email address, and "domain" when the user's email isn't a generic one.