Add slug to Category
This will allow categories to be updated from the admin interface via their slugs as opposed to their DB IDs, which are not guaranteed to always be the same, especially when testing both the API and admin interface locally. The specs for the admin interface currently communicate with an instance of the API running locally on the same machine as the admin interface.